+++ spec/kettle/changelog_spec.rb
@@ -5,8 +5,9 @@ RSpec.describe Kettle::Changelog do
     expect(Kettle::Changelog::VERSION).not_to be_nil
   end
 
-  it "exposes the changelog CLI and entry adder" do
+  it "exposes the changelog CLI and entry helpers" do
     expect(described_class::CLI).to be < Object
     expect(described_class::EntryAdder).to be < Object
+    expect(described_class::KeyedEntryUpserter).to be < Object
   end
 end